PR feedback: moved EventHubPartitionKeyResolver under /test
(cherry picked from commit 85c926baad0022500d13f4c50912c90b0d1d4cf2)
